+++ ArchitectureAnalyzer.Tests/AnalyzerRules/Rules/MustImplementRuleTests.cs
@@ -1,12 +1,14 @@
 ﻿using ArchitectureAnalyzer.AnalyzerRules.Rules;
 using ArchitectureAnalyzer.Tests.AnalyzerRules.Matchers;
 using ArchitectureAnalyzer.Tests.Helper;
+using Microsoft.CodeAnalysis;
 
 namespace ArchitectureAnalyzer.Tests.AnalyzerRules.Rules;
 
 public sealed class MustImplementRuleTests
 {
 	private readonly MustImplementRule _sut = new();
+	private readonly Compilation _compilation = default!;
 
 	[Fact]
 	public void DoesNotViolate_WhenForbidden_WhenInterfaceIsNotImplemented()
@@ -22,7 +24,7 @@ public sealed class MustImplementRuleTests
 		_sut.ForTypes = forTypes;
 
 		// Act
-		var result = _sut.Evaluate(symbol);
+		var result = _sut.Evaluate(symbol, _compilation);
 
 		// Assert
 		result.Should().BeNull();
@@ -42,7 +44,7 @@ public sealed class MustImplementRuleTests
 		_sut.ForTypes = forTypes;
 
 		// Act
-		var result = _sut.Evaluate(symbol);
+		var result = _sut.Evaluate(symbol, _compilation);
 
 		// Assert
 		result.Should().BeNull();
@@ -61,7 +63,7 @@ public sealed class MustImplementRuleTests
 		_sut.ForTypes = forTypes;
 
 		// Act
-		var result = _sut.Evaluate(symbol);
+		var result = _sut.Evaluate(symbol, _compilation);
 
 		// Assert
 		result.Should().BeNull();
@@ -80,7 +82,7 @@ public sealed class MustImplementRuleTests
 		_sut.ForTypes = forTypes;
 
 		// Act
-		var result = _sut.Evaluate(symbol);
+		var result = _sut.Evaluate(symbol, _compilation);
 
 		// Assert
 		result.Should().BeNull();
@@ -100,7 +102,7 @@ public sealed class MustImplementRuleTests
 		_sut.ForTypes = forTypes;
 
 		// Act
-		var result = _sut.Evaluate(symbol);
+		var result = _sut.Evaluate(symbol, _compilation);
 
 		// Assert
 		result.Should().NotBeNull();
@@ -119,7 +121,7 @@ public sealed class MustImplementRuleTests
 		_sut.ForTypes = forTypes;
 
 		// Act
-		var result = _sut.Evaluate(symbol);
+		var result = _sut.Evaluate(symbol, _compilation);
 
 		// Assert
 		result.Should().NotBeNull();