+++ ArchitectureAnalyzer/AnalyzerRules/Rules/Rule.cs
@@ -5,28 +5,29 @@ namespace ArchitectureAnalyzer.AnalyzerRules.Rules;
 
 internal abstract class Rule
 {
-	public Matcher ForTypes { get; set; } = default!;
-	public bool Forbidden { get; set; }
+	// ReSharper disable once UnusedAutoPropertyAccessor.Global
 	public string Description { get; set; } = default!;
+	public bool Forbidden { get; set; }
+	public Matcher ForTypes { get; set; } = default!;
 
 	protected abstract DiagnosticDescriptor Descriptor { get; }
 
-	public Diagnostic? Evaluate(INamedTypeSymbol symbol)
+	public Diagnostic? Evaluate(INamedTypeSymbol symbol, Compilation compilation)
 	{
 		if (!ForTypes.Matches(symbol))
 			return null;
 
-		return EvaluateInternal(symbol);
+		return EvaluateInternal(symbol, compilation);
 	}
-
-	protected abstract Diagnostic? EvaluateInternal(INamedTypeSymbol symbol);
 
 	protected Diagnostic CreateDiagnostic(INamedTypeSymbol symbol, params string[] messageArgs)
 	{
 		var location = symbol.Locations.First();
 		var additionalLocations = symbol.Locations.Skip(1);
-		
+
 		// ReSharper disable once CoVariantArrayConversion
 		return Diagnostic.Create(Descriptor, location, additionalLocations, messageArgs);
 	}
+
+	protected abstract Diagnostic? EvaluateInternal(INamedTypeSymbol symbol, Compilation compilation);
 }